Initial Checkin
This is the very beginning of a work in progress.  It currently allows
creating a sparse vector and setting values in it.  I’ve limited the
NODE_SIZE to just 4, so I can hand code all the functionality prior to
creating a source generator for larger node sizes (like 32).  Big
things to do before that happens
- remove items when set to zero value
- add push() methods to append values to the end of the vector (and so
to increase the size of the vector)
- add functional methods map, filter, etc.